+++ init.py
@@ -1,0 +1,18 @@
+import pygame
+import numpy as np
+
+# Grootte van het bord
+field = HEIGHT, WIDTH = (20, 10)
+# Maakt een 2d lijst gevuld met 0
+board = np.zeros(field, dtype=int)
+
+SQUARE_SIZE = 30
+window_shape = window_width, window_height = (
+    (WIDTH + 2 + 9) * SQUARE_SIZE, (HEIGHT + 1) * SQUARE_SIZE)
+
+pygame.display.init()
+pygame.font.init()
+screen = pygame.display.set_mode(window_shape)
+pygame.display.set_caption('Tetris')
+clock = pygame.time.Clock()
+font = pygame.font.Font("Redkost Comic.otf", 32)
